refactor: refactor functions to interactive.go
